refactor(onboarding): split onboardingaggregate event records into separate files
Move all eight event records (OnboardingStarted, TopicAsked,
UserTurnRecorded, AssistantTurnRecorded, AnswerCaptured,
ClarificationRequested, PlanLinkedToUser, OnboardingCompleted) each
into their own file to satisfy the one-type-per-file convention.
Remove the SA1402/SA1649 pragma suppressions from OnboardingAggregate.cs.
